Все недочеты были учтены и исправлены
